+++ src/cli/main.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { buildProgram } from "./index.js";
+
+buildProgram().parse(process.argv);